+++ KGEN/utils/vscode-mojo/CHANGELOG.md
@@ -4,6 +4,25 @@
 
 ### Added
 
+- The Mojo LSP server will now generate fixits for populating empty
+  documentation strings:
+
+  ```mojo
+  fn foo(arg: Int):
+    """""" # Unexpected empty documentation string
+  ```
+
+  Applying the fixit from above will generate:
+
+  ```mojo
+  fn foo(arg: Int):
+    """[summary].
+
+    Args:
+        arg: [description].
+    """
+  ```
+
 ### Changed
 
 ### Fixed